refactor: Move RateCalculator to utilities
