+++ src/components/utilities/RateCalculator.js
